backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pypdf==5.0.1
sentence-transformers==3.1.1
numpy==1.26.4
httpx==0.27.2
python-dotenv==1.0.1
